CD-7792 updated svg code to file icons
