+++ server/sonar-web/src/main/js/api/ai-codefix.ts
@@ -17,7 +17,21 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-import { postJSONBody } from '../helpers/request';
+import {
+  postJSONBody,
+  getCSRFToken,
+  parseJSON,
+  post,
+  checkStatus,
+} from '../helpers/request';
+import { getBaseUrl } from '../helpers/system';
+
+export interface CodefixFixedFileResponse {
+  jobId: string;
+  fixedFileContent: string;
+}
+
+const CODEFIX_BASE = '/_codescan/codefix';
 
 export function queueCodeFix(data: {
     organizationKey: string;
@@ -25,4 +39,29 @@ export function queueCodeFix(data: {
     issueKey: string;
 }): Promise<void> {
   return postJSONBody('/_codescan/codefix/queue', data);
+}
+
+/**
+ * Fetches the fixed file content for an issue from the internal codefix API.
+ * Request is sent with same-origin credentials (session cookie) so the backend
+ * can authenticate the user. The backend must restrict this API to logged-in
+ * users and, when proxying to CodeScan, forward the request with the SQ token
+ * so it is not callable from outside with a raw token.
+ */
+export function getCodefixFixedFile(issueKey: string): Promise<CodefixFixedFileResponse> {
+  const url = `${getBaseUrl()}${CODEFIX_BASE}/fixed-file?issueKey=${encodeURIComponent(issueKey)}`;
+  return fetch(url, {
+    method: 'GET',
+    credentials: 'same-origin',
+    headers: {
+      Accept: 'application/json',
+      ...getCSRFToken(),
+    },
+  })
+    .then(checkStatus)
+    .then(parseJSON);
+}
+
+export function createCodefixPr(jobId: string): Promise<void> {
+  return post(`${CODEFIX_BASE}/create-pr/${encodeURIComponent(jobId)}`);
 }
